hjghj

+++ SPG-Financas/src/main/java/br/com/empresaSGP/SPG_Financas/entity/ContasReceberEntity.java
@@ -32,7 +32,7 @@ public class ContasReceberEntity implements Serializable {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
+	private Long id;
 	
 	@NotNull
 	private long idFornecedores;
@@ -87,11 +87,13 @@ public class ContasReceberEntity implements Serializable {
 		this.tipoconta = tipoconta;
 	}
 
-	public int getId() {
+	
+
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
